Add diablo chruch arrival vault
adds a church based on the tristram cathedral from the first
diablo game. Hell has desecrated the altar, but a player must
first contend with the Imp guarding it if they dare. a hatch is
provided so you can worship and run or hatch if you get an
unlucky teleport. To the sides of the church there is a crypt and
through the windows the player can see the lust forest they leave
behind. There's also a baptism font (usually dry) because why not.